Add linked data entities.
`static_attributes` may be supplied with an additional `link` data element when provisioning an IoT Agent to ensure that active attributes from the provisioned IoT Device may be maintained in parallel with a linked data entity . Take for example a temperature gauge placed within a building. The **Device** data model literally represents the IoT device itself, but the `temperature` attribute also needs to be shared with the **Building** entity. A `link` between them can be provisioned